ec2_python_scripts/warc_file_extractor_script/requirements.txt
newspaper3k
warcio
lxml_html_clean